+++ public/admin/_app/immutable/nodes/24.Bd0Jy3U3.js
@@ -1,0 +1,1 @@
+import{t as e}from"../chunks/Cy2AE35z.js";export{e as component};